Move CustomizableAttachment.tsx at the correct place
