+++ react/javascript/src/components/customizable/CustomizableAttachment.tsx
@@ -1,5 +1,5 @@
-import { IAttachmentProps } from '../../../src/components/gherkin/Attachment'
-import MessageToComponentMappingContext from '../../../src/contexts/MessageToComponentMappingContext'
+import { IAttachmentProps } from '../gherkin/Attachment'
+import MessageToComponentMappingContext from '../../contexts/MessageToComponentMappingContext'
 import React from 'react'
 
 const CustomizableAttachment: React.FunctionComponent<IAttachmentProps> = ({
