+++ react/javascript/test/components/gherkin/StepTest.tsx
@@ -11,7 +11,9 @@ import GherkinQueryContext from '../../../src/contexts/GherkinQueryContext'
 import { Query as GherkinQuery } from '@cucumber/gherkin'
 import CucumberQueryContext from '../../../src/contexts/CucumberQueryContext'
 import { IAttachmentProps } from '../../../src/components/gherkin/Attachment'
-import MessageToComponentMappingContext from '../../../src/contexts/MessageToComponentMappingContext'
+import MessageToComponentMappingContext, {
+  defaultMessageToComponentMapping,
+} from '../../../src/contexts/MessageToComponentMappingContext'
 
 describe('<Step>', () => {
   const dom = new JSDOM(
@@ -94,6 +96,11 @@ describe('<Step>', () => {
     }) => {
       return <div>{attachment.body.toUpperCase()}</div>
     }
+    const messageToComponentMapping = {
+      ...defaultMessageToComponentMapping,
+      ...{ attachment: CapitalizedAttachment },
+    }
+
     const app = (
       <GherkinQueryContext.Provider value={new StubGherkinQuery()}>
         <UriContext.Provider value={'some.feature'}>
@@ -101,7 +108,7 @@ describe('<Step>', () => {
             value={new StubCucumberQueryWithTextAttachment()}
           >
             <MessageToComponentMappingContext.Provider
-              value={{ attachment: CapitalizedAttachment }}
+              value={messageToComponentMapping}
             >
               <Step
                 step={step}
